Проектная работа - исправления
