+++ src/constants/api.ts
@@ -1,1 +1,1 @@
-export const url = 'https://norma.nomoreparties.space/api/ingredients';
+export const url = 'https://norma.nomoreparties.space/api';
